wip: cleanup code
