--- deploy/helm/templates/configmap.yaml
@@ -1,37 +1,0 @@
-apiVersion: v1
-kind: ConfigMap
-metadata:
-  name: {{ template "calert.fullname" . }}-config
-  labels:
-    app: {{ template "calert.name" . }}
-    chart: {{ template "calert.chart" . }}
-    heritage: {{ .Release.Service }}
-    release: {{ .Release.Name }}
-data:
-  message.tmpl: {{- toYaml .Values.template_file | indent 4 }}
-  config.toml: |
-    # All timeouts and durations are in milliseconds.
-
-    [server]
-    address = {{ .Values.configmap.server.address | quote }}
-    socket = {{ .Values.configmap.server.socket | quote }}
-    name = {{ .Values.configmap.server.name | quote }}
-
-    # WARNING If these timeouts are less than 1s,
-    # the server connection breaks.
-    read_timeout={{ .Values.configmap.server.read_timeout }}
-    write_timeout={{ .Values.configmap.server.write_timeout }}
-    keepalive_timeout={{ .Values.configmap.server.keepalive_timeout }}
-
-    [app]
-    template_file = {{ .Values.configmap.app.template_file | quote }}
-    max_size = {{ .Values.configmap.app.max_size }}
-
-    [app.http_client]
-    max_idle_conns = {{ .Values.configmap.app.http_client.max_idle_conns }}
-    request_timeout = {{ .Values.configmap.app.http_client.request_timeout }}
-    proxy_url = {{ .Values.configmap.app.http_client.proxy_url | quote }}
-
-    {{- with .Values.configmap.rooms }}
-    {{ tpl . $ | indent 4 }}
-    {{- end }}
