--- deploy/helm/templates/service.yaml
@@ -1,15 +1,0 @@
-apiVersion: v1
-kind: Service
-metadata:
-  name: {{ include "calert.fullname" . }}
-  labels:
-{{ include "calert.labels" . | indent 4 }}
-spec:
-  type: {{ .Values.service.type }}
-  ports:
-    - port: {{ .Values.service.port }}
-      targetPort: {{ .Values.service.port }}
-      protocol: TCP
-  selector:
-    app.kubernetes.io/name: {{ include "calert.name" . }}
-    app.kubernetes.io/instance: {{ .Release.Name }}
