--- deploy/helm/values.yaml
@@ -1,69 +1,0 @@
-# Default values for calert.
-# This is a YAML-formatted file.
-# Declare variables to be passed into your templates.
-
-configmap:
-    # All timeouts and durations are in milliseconds.
-  server:
-    address: ":6000"
-    socket: "/tmp/calert.sock"
-    name: "calert"
-    # WARNING If these timeouts are less than 1s, the server connection breaks.
-    read_timeout: "8000"
-    write_timeout: "8000"
-    keepalive_timeout: "300000"
-  app:
-    template_file: "/etc/calert/message.tmpl"
-    max_size: 4000
-    http_client:
-      max_idle_conns: "100"
-      request_timeout: "8000"
-      proxy_url: "http://internal-squid-proxy.com:3128"
-  rooms: |
-
-    [app.chat.alertManagerTestRoom]
-    notification_url = "https://chat.googleapis.com/v1/spaces/xxx/messages?key=abc-xyz&token=token-unique-key%3D"
-
-    [app.chat.awesomeRoomTwo]
-    notification_url = "https://chat.googleapis.com/v1/spaces/xxx/messages?key=abc-xyz&token=token-unique-key%3D"
-
-template_file: |
-  *{{ .Labels.alertname | Title }} - {{.Status | Title }} ({{.Labels.severity | toUpper }})*
-  {{ range .Annotations.SortedPairs -}}
-  {{ .Name }}: {{ .Value}}
-  {{ end -}}
-
-replicaCount: 1
-
-image:
-  repository: mrkaran/calert
-  tag: v1.2.0
-  pullPolicy: Always
-
-command:
-- "./calert"
-- "--config.file=/etc/calert/config.toml"
-
-nameOverride: ""
-fullnameOverride: ""
-
-service:
-  type: ClusterIP
-  port: 6000
-
-ingress:
-  enabled: false
-
-resources:
-  limits:
-    cpu: 20m
-    memory: 48Mi
-  requests:
-    cpu: 5m
-    memory: 24Mi
-
-nodeSelector: {}
-
-tolerations: []
-
-affinity: {}
