--- deploy/kustomize/base/deployments/calert.yaml
@@ -1,27 +1,0 @@
-apiVersion: apps/v1
-kind: Deployment
-metadata:
-  name: app
-spec:
-  replicas: 1
-  selector:
-    matchLabels:
-      service: app
-  template:
-    metadata:
-      labels:
-        service: app
-    spec:
-      containers:
-      - name: calert
-        image: mrkaran/calert:1.0.0
-        ports:
-        - containerPort: 6000
-        volumeMounts:
-        - mountPath: /etc/calert/config.toml
-          name: config-dir
-          subPath: config.toml
-      volumes:
-        - name: config-dir
-          configMap:
-            name: app-config
