--- deploy/kustomize/base/services/calert.yaml
@@ -1,11 +1,0 @@
-apiVersion: v1
-kind: Service
-metadata:
-  name: service
-spec:
-  ports:
-  - protocol: TCP
-    port: 6000
-  selector:
-    tier: api
-  type: NodePort
