--- deploy/kustomize/overlays/prod/kustomization.yaml
@@ -1,20 +1,0 @@
-apiVersion: kustomize.config.k8s.io/v1beta1
-kind: Kustomization
-namespace: calert
-commonLabels:
-  env: prod
-  group: calert
-resources:
-  - namespace.yaml
-  - ../../base
-patchesStrategicMerge:
-- patches/resource-limits.yaml
-- patches/add-labels.yaml
-
-images:
-  - name: mrkaran/calert
-    newTag: 1.0.0-stable
-configMapGenerator:
-- name: app-config
-  files:
-    - raw/config.toml
