--- deploy/kustomize/overlays/prod/patches/add-labels.yaml
@@ -1,9 +1,0 @@
-apiVersion: apps/v1
-kind: Deployment
-metadata:
-  name: app
-spec:
-  template:
-    metadata:
-      labels:
-        tier: api
