--- deploy/kustomize/overlays/prod/patches/resource-limits.yaml
@@ -1,16 +1,0 @@
-apiVersion: apps/v1
-kind: Deployment
-metadata:
-  name: app
-spec:
-  template:
-    spec:
-      containers:
-        - name: calert
-          resources:
-            requests:
-              memory: "200Mi"
-              cpu: "250m"
-            limits:
-              memory: "400Mi"
-              cpu: "500m"
